requirements.txt
flask
flask-cors
flask-migrate
flask-sqlalchemy
flask-mail
python-dotenv
psycopg2-binary
flask-jwt-extended
bcrypt
gunicorn
GitPython